ml-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
numpy==1.26.4
tensorflow==2.17.0
scikit-learn==1.5.2
pandas==2.2.3
google-cloud-aiplatform==1.67.1
python-dateutil==2.9.0.post0
google-cloud-secret-manager==2.20.0